+++ client/src/index.js
@@ -1,9 +1,9 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
-import store from "./app/store";
+import store from "./reducers";
 import { Provider } from "react-redux";
-import { Game } from "./components/Game";
+import { Game } from "./contaniers/Game";
 
 ReactDOM.render(
   <React.StrictMode>
