+++ client/src/reducers/index.js
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import gameReducer from "../reducers/gameSlice";
+import gameReducer from "./gameSlice";
 
 export default configureStore({
   reducer: {
